Use assembler implementation of TLS initialization with clang
Clang treats register gp as a callee-saved register. In the C
implementation of the TLS initialization code, it detect the
modification to the register and therefore saves it on the stack thereby
undoing the TLS initialization.

Work around this issue by implementing the TLS initialization in
assembler and not telling clang that gp has been modified.

Signed-off-by: Lukas Auer <lukas.auer@aisec.fraunhofer.de>