ioctl: introduce ioctl wrapper
Cheri changes the calling convention for variadic functions, passing
all variadic params on stack. The autogenerated glibc syscall wrapper
for ioctl does not expect that and does, hence, not work for cheri.
The wrapper pulls the variadic third ioctl argument from the stack and
passes it correctly in a register.