Bounds relocation of object
When re-relocating rtld the phdr field of the DSO link_map is a read-only capability, based on the asm lookup of the __ehdr_start at the beginning of the elf header. Due to the pc relative loading, the permissions are set to read-only and execute. The phdr capability is used to initialize the data capability used in __cheri_init_globals, which needs to enable writes, too. Redirect capability through DDC.